NET BS 7.0: Droid - Fixed the comments, Removed unused files, Fixed the snippets. - Renamed files, with proper naming conventions.
