+++ BarcodeSDK.NET.Droid.Example/AndroidManifest.xml
@@ -6,7 +6,7 @@
 	<uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 	<uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 	
-	<!-- Enables downloading of images from URLs in RTU v2 -->
+	<!-- Enables downloading of images from URLs in RTU UI -->
 	 <uses-permission android:name="android.permission.INTERNET" />
 	<application android:label="BarcodeSDK.NET" android:theme="@style/AppTheme" />
 </manifest>
