+++ BarcodeSDK.NET.Droid.Example/Snippets/UserGuidance.cs
@@ -12,7 +12,7 @@ public static partial class Snippets
             // Create the default configuration object.
             var config = new BarcodeScannerScreenConfiguration();
             
-            // Hide/unhide the user guidance.
+            // Hide/Show the user guidance.
             config.UserGuidance.Visible = true;
 
             // Configure the title.
